refactor: changed general project tree
