+++ .github/workflows/ci.yml
@@ -27,7 +27,7 @@ jobs:
           if [ -f requirements_dev.txt ]; then pip install -r requirements_dev.txt; fi
       - name: Test with pytest
         run: |
-          pytest --cov src tests/ --cov-fail-under=100 --cov-report xml
+          pytest --cov . tests/
       - name: Upload to CodeCoverage
         uses: codecov/codecov-action@v2
         with:
