+++ module/shared.py
@@ -1,0 +1,4 @@
+from json import load
+
+with open("data/welcome.json", "r") as f:
+    welcome = load(f)
