+++ requirements.txt
@@ -1,1 +1,1 @@
-python-telegram-bot == 20.0a0
+python-telegram-bot==13.8.1
